GenerateRequest and LiteralDataRequest should not include execution
