+++ core/src/main/java/com/redhat/lightblue/client/request/AbstractLightblueDataRequest.java
@@ -9,8 +9,6 @@ import com.redhat.lightblue.client.Operation;
 
 public abstract class AbstractLightblueDataRequest extends AbstractLightblueRequest implements LightblueRequest {
 
-    private Execution execution = null;
-
     public AbstractLightblueDataRequest(String entityName, String entityVersion) {
         super(entityName, entityVersion);
     }
@@ -46,16 +44,6 @@ public abstract class AbstractLightblueDataRequest extends AbstractLightblueRequ
                 node.set("maxResults", JsonNodeFactory.instance.numberNode(maxResults));
             }
         }
-    }
-
-    public AbstractLightblueDataRequest execution(Execution execution) {
-        this.execution = execution;
-        return this;
-    }
-
-    protected void appendExecutionToJson(ObjectNode node) {
-        if (execution != null)
-            node.set("execution", execution.toJson());
     }
 
     @Override
